config: forbid attempt to join read-only replicaset
If the given instance has no existing snapshot, it means that it is not
registered in the replicaset and will attempt to join as a new replica.
It has no chance to succeed if all the instances in the replicaset are
in the read-only mode: neither of them can register the new instance.

Report an error and fail the startup in the case to make the situation
clear.

Part of #8810

NO_DOC=the old behavior was not released, the documentation request will
       be registered manually
NO_CHANGELOG=see NO_DOC