box: switch to normal streaming compression API of ZSTD
For writing xlogs, we use buffer-less compression API of ZSTD, which is
deprecated in version 1.5.5 in favor of the normal streaming API [1]:
> This API is deprecated, and will be removed in a future version. It
> allows streaming (de)compression with user allocated buffers. However,
> it is hard to use, and not as well tested as the rest of our API.
>
> Please use the normal streaming API instead: ZSTD_compressStream2, and
> ZSTD_decompressStream.

Refactor `xlog_tx_write_zstd` to use the normal streaming API, as
suggested. We decided to flush (i.e., do `obuf_alloc` plus `crc32_calc`)
the ZSTD output buffer unconditionally (as opposed to waiting for it to
fill up), since in our case flushing is cheap and we do not risk getting
stuck waiting for the buffer to be filled completely [2].

1. https://github.com/facebook/zstd/blob/25822342be59d831bad65426ae51f5cc22157b09/lib/zstd.h#L2822-L2834
2. Imagine a corner case, where the output buffer has 1 byte of space left,
but ZSTD needs to write at least two bytes (e..g., for compressing an
'aaaa' sequence it needs to write '4a').

NO_CHANGELOG=refactoring
NO_DOC=refactoring
NO_TEST=refactoring