ci/cd: drop distro-specific packaging workflows
Tarantool 3.0.0 is not going to be distributed as a set of packages for
specific distros. Instead, it is planned to have a few `deb` and `rpm`
packages with a statically compiled Tarantool binary inside, which will
work on almost any distro. Also, we drop building packages with GC32.
Now only GC64 packages will be shipped.

Static build packaging was added at #8771.

NO_DOC=ci
NO_TEST=ci