config: restrict iproto.advertise.client validation
The following syntax variants are fobidden now:

* user@
* user:pass@
* user@host:port
* user:pass@host:port

Only host:port is allowed (inet URI or unix socket URI).

The idea is that we shouldn't distribute login/password information
using the configuration, so it is better to explicitly forbid such
usage of the option.

Closes #8810

NO_DOC=the old behavior was not released, the documentation request will
       be registered manually
NO_CHANGELOG=see NO_DOC