asan: adapt misc stats test for ASAN
When SMALL_MALLOC_IMPL is defined and ASAN-friendly allocators are used
the arena allocator is not used at all as we not allocate memory
directly from there. And other ASAN-friendly allocators are not allocate
from it too. Thus box.slab.info().arena_size == 0. Same for usage
of runtime arena box.runtime.info().used.

Also usage with ASAN-friendly lsregion is a bit different as it does
not account for size of alignment padding. Thus we need to adapt
box.stat.vinyl().memory.level0 tests. Approach is to check for lower
and upper limit instead of checking for exact values.

Part of #7327

NO_DOC=test changes
NO_CHANGELOG=test changes