func: introduce new option trigger
This option will be used as an event name for this function - when its
tuple is replaced in space _func, the function will be inserted to the
trigger registry. Option trigger specifies the name of event in which
the trigger will be inserted, also one can pass an array of strings
to set a trigger in several events.

Along the way, fix inappropriate error thrown when `param_list` options
of a function contains object of invalid type (only string is expected):
ER_FIELD_TYPE, which was used there, is about a tuple, not array.

Part of #8663

NO_CHANGELOG=later
NO_DOC=later