func: introduce new option trigger
This option will be used as an event name for this function - when its
tuple is replaced in space _func, the function will be inserted to the
trigger registry. Option trigger specifies the name of event in which
the trigger will be inserted, also one can pass an array of strings
to set a trigger in several events.

Part of #8663

NO_CHANGELOG=later
NO_DOC=later