config: test that no box.cfg() options are missed
A few options were added into box.cfg during development of the
declarative configuration module. As result the instance config schema
misses them. We're going to add these new options in #8861.

However, it would be nice to catch such situations automatically. The
new test case verifies it.

If a new box.cfg() option will be added or a default value of an
existing option will be changed, the test case will highlight it to let
a human decide what to do with the declarative config.

Part of #8861

NO_DOC=testing change
NO_CHANGELOG=see NO_DOC