box: deprecate box.session.push
Closes #8802

@TarantoolBot document
Title: Deprecate box.session.push

The `box.session.push` Lua function and `box_session_push` C API
function are deprecated starting from Tarantool 3.0. Calling any
of these functions for the first time results in printing a warning
message to the log.

The new compat module option `box_session_push_deprecation` was
introduced to control whether the functions are still available.
With the old behavior, which is the default in Tarantool 3.0,
`box.session.push` is still available. With the new behavior,
any attempt to use it raises an exception.

(Please create https://tarantool.io/compat/box_session_push_deprecation)

We are planning to switch the compat option to the new behavior
starting from Tarantool 4.0 with the ability to revert to the
old behavior. Starting from Tarantool 5.0 we are planning to
drop `box.session.push` completely.