lua/popen: introduce inherit_fds option for popen.new
Closes #8926

@TarantoolBot document
Title: Document `inherit_fds` option of `popen.new`

The new option takes an array of file descriptor numbers that should be
left open in the child process if the `close_fds` flag is set. If the
`close_fds` flag isn't set, the option has no effect because in this
case none of the parent's file descriptors are closed anyway.

The option may be useful for establishing a communication channel
between a Tarantool instance and a process spawned by it with
`popen.new`, for example, using `socket.socketpair`.