box: forbid same names for check constraints and foreign keys
It is possible to create a set of check constraints and foreign keys for
each field and for entire tuple. By design for each field/tuple the check
constraints must have unique names, the same is true for foreign keys. But
it is possible to create a check constraint and a foreign key with the same
name. This introduces difficulties in SQL implementation. On the other hand
in real life it is hard to imagine a user that wants to create both check
and foreign constraints for a field, especially with the same names. As for
entire tuple it seems less obvious, but still it seems that there will be
no harm if we forbid same names.

Closes #7650

NO_DOC=minor