iproto: use xobuf_alloc for output buffer allocations
The buffer is unlimited so allocations from it never fail. Even if
decide to add a hard limit some day, it's better to do it somewhere in
one place to simplify testing so the dropped code isn't going to be
needed.

NO_DOC=code cleanup
NO_TEST=code cleanup
NO_CHANGELOG=code cleanup